Update README.md
This project focuses on the application of python tools to analyze the different factors affecting ticket price variation in the given dataset. Question were answered through EDA and a viable predicion model was developed. This is aimed at guiding indivuals looking to book a flight in the future on decision making and could also help airlines improve on their services through an "Ease My Trip" website.